+++ cmd/cleanup.go
@@ -9,9 +9,10 @@ import (
 )
 
 var (
-	allFlag    bool
-	orphanFlag bool
-	daysOption int
+	allFlag       bool
+	orphanFlag    bool
+	minicondaFlag bool
+	daysOption    int
 )
 
 var cleanupCmd = &cobra.Command{
@@ -23,7 +24,7 @@ After cleanup, they will not be available anymore.`,
 		if common.DebugFlag {
 			defer common.Stopwatch("Env cleanup lasted").Report()
 		}
-		err := conda.Cleanup(daysOption, dryFlag, orphanFlag, allFlag)
+		err := conda.Cleanup(daysOption, dryFlag, orphanFlag, allFlag, minicondaFlag)
 		if err != nil {
 			pretty.Exit(1, "Error: %v", err)
 		}
@@ -35,6 +36,7 @@ func init() {
 	envCmd.AddCommand(cleanupCmd)
 	cleanupCmd.Flags().BoolVarP(&dryFlag, "dryrun", "d", false, "Don't delete environments, just show what would happen.")
 	cleanupCmd.Flags().BoolVarP(&orphanFlag, "orphans", "o", false, "Cleanup orphan, unreachable enviroments.")
+	cleanupCmd.Flags().BoolVarP(&minicondaFlag, "miniconda", "m", false, "Remove miniconda3 installation (replaced by micromamba).")
 	cleanupCmd.Flags().BoolVarP(&allFlag, "all", "a", false, "Cleanup all enviroments.")
 	cleanupCmd.Flags().IntVarP(&daysOption, "days", "", 30, "What is the limit in days to keep environments for (deletes environments older than this).")
 }